Replace existing copy-pasted java.nio classes with code generated from templates
