--- javalib/src/main/scala/java/nio/MappedByteBufferFloatView.scala
@@ -1,115 +1,0 @@
-package java.nio
-
-// Based on the code ported from Scala.js,
-// see HeapByteBufferFloatView.scala
-private[nio] final class MappedByteBufferFloatView private (
-    _capacity: Int,
-    override private[nio] val _mappedData: MappedByteBufferData,
-    override private[nio] val _byteArrayOffset: Int,
-    _initialPosition: Int,
-    _initialLimit: Int,
-    _readOnly: Boolean,
-    override private[nio] val isBigEndian: Boolean
-) extends FloatBuffer(_capacity, null, null, -1) {
-
-  position(_initialPosition)
-  limit(_initialLimit)
-
-  private def genBuffer = GenBuffer[FloatBuffer](this)
-  protected def genMappedBufferView = GenMappedBufferView[FloatBuffer](this)
-  private lazy val byteArrayBits = genMappedBufferView.byteArrayBits
-  private[this] implicit def newMappedFloatBufferView
-      : GenMappedBufferView.NewMappedBufferView[FloatBuffer] =
-    MappedByteBufferFloatView.NewMappedByteBufferFloatView
-
-  def isReadOnly(): Boolean = _readOnly
-
-  def isDirect(): Boolean = true
-
-  @noinline
-  def slice(): FloatBuffer =
-    genMappedBufferView.generic_slice()
-
-  @noinline
-  def duplicate(): FloatBuffer =
-    genMappedBufferView.generic_duplicate()
-
-  @noinline
-  def asReadOnlyBuffer(): FloatBuffer =
-    genMappedBufferView.generic_asReadOnlyBuffer()
-
-  @noinline
-  def get(): Float =
-    genBuffer.generic_get()
-
-  @noinline
-  def put(c: Float): FloatBuffer =
-    genBuffer.generic_put(c)
-
-  @noinline
-  def get(index: Int): Float =
-    genBuffer.generic_get(index)
-
-  @noinline
-  def put(index: Int, c: Float): FloatBuffer =
-    genBuffer.generic_put(index, c)
-
-  @noinline
-  override def get(dst: Array[Float], offset: Int, length: Int): FloatBuffer =
-    genBuffer.generic_get(dst, offset, length)
-
-  @noinline
-  override def put(src: Array[Float], offset: Int, length: Int): FloatBuffer =
-    genBuffer.generic_put(src, offset, length)
-
-  @noinline
-  def compact(): FloatBuffer =
-    genMappedBufferView.generic_compact()
-
-  @noinline
-  def order(): ByteOrder =
-    genMappedBufferView.generic_order()
-
-  // Private API
-
-  @inline
-  private[nio] def load(index: Int): Float =
-    byteArrayBits.loadFloat(index)
-
-  @inline
-  private[nio] def store(index: Int, elem: Float): Unit =
-    byteArrayBits.storeFloat(index, elem)
-}
-
-private[nio] object MappedByteBufferFloatView {
-  private[nio] implicit object NewMappedByteBufferFloatView
-      extends GenMappedBufferView.NewMappedBufferView[FloatBuffer] {
-    def bytesPerElem: Int = 4
-
-    def apply(
-        capacity: Int,
-        mappedData: MappedByteBufferData,
-        byteArrayOffset: Int,
-        initialPosition: Int,
-        initialLimit: Int,
-        readOnly: Boolean,
-        isBigEndian: Boolean
-    ): FloatBuffer = {
-      new MappedByteBufferFloatView(
-        capacity,
-        mappedData,
-        byteArrayOffset,
-        initialPosition,
-        initialLimit,
-        readOnly,
-        isBigEndian
-      )
-    }
-  }
-
-  @inline
-  private[nio] def fromMappedByteBuffer(
-      byteBuffer: MappedByteBuffer
-  ): FloatBuffer =
-    GenMappedBufferView.generic_fromMappedByteBuffer(byteBuffer)
-}
